MD5

+++ client/src/main/java/com/management/client/service/impl/UserServiceImpl.java
@@ -2,6 +2,7 @@ package com.management.client.service.impl;
 
 import com.management.client.dao.UserDao;
 import com.management.client.service.UserService;
+import com.management.client.unit.MD5;
 import com.management.client.vo.common.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,12 +11,10 @@ import java.util.Date;
 import java.util.List;
 
 /**
- *
  * UserServiceImpl
  *
  * @author 严虹钱
- * @since  2022/3/25
- *
+ * @since 2022/3/25
  */
 
 @Service
@@ -27,6 +26,8 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User login(User user) {
+        String passWord = MD5.encrypt2ToMD5(user.getPassword());
+        user.setPassword(passWord);
         User result = userDao.login(user);
         if (result == null) {
             throw new IllegalArgumentException("账号或密码错误！");
@@ -36,6 +37,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Integer id) {
+        if (id == 1) {
+            throw new IllegalArgumentException("此为系统初始管理员无法删除！");
+        }
         userDao.delete(id);
     }
 
@@ -43,6 +47,8 @@ public class UserServiceImpl implements UserService {
     public void insert(User user) {
         this.chekEmpty(user);
         check(user);
+        String passWord = MD5.encrypt2ToMD5(user.getPassword());
+        user.setPassword(passWord);
         user.setCreateTime(new Date());
         String workNumber = System.currentTimeMillis() + "";
         user.setWorkNumber(workNumber);
@@ -59,10 +65,17 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void update(User user) {
+        if (user.getId() == 1) {
+            throw new IllegalArgumentException("此为系统初始管理员无法修改！");
+        }
+        User old = userDao.getUserById(user.getId());
+        if (!old.getPassword().equals(user.getPassword())) {
+            String passWord = MD5.encrypt2ToMD5(user.getPassword());
+            user.setPassword(passWord);
+        }
         this.chekEmpty(user);
-        User OldUser = userDao.getUserById(user.getId());
         List<User> users = userDao.getAllUser(new User());
-        if (users.stream().anyMatch(User -> User.getAccount().equals(user.getAccount())) && !user.getAccount().equals(OldUser.getAccount())) {
+        if (users.stream().anyMatch(User -> User.getAccount().equals(user.getAccount())) && !user.getAccount().equals(old.getAccount())) {
             throw new IllegalArgumentException("该账号已经存在请重新输入");
         }
         userDao.update(user);
